LPD-57772 Extract chooseFileFromDocumentLibrary to reuse it
